Fix error while allocating shared memory.
If num_init_children * max_pool is huge, pool_coninfo_size() failed to
calculate the size of required shared memory because the data type
used for the variable to store shared memory size was int and it
overflows for too large shared memory size. To fix this change
pool_coninfo_size() to size_t and use appropriate size_t for the
variable used within the function.

Also fix to use proper format qualifier in
pool_shared_memory_create().

+++ src/include/context/pool_process_context.h
@@ -6,7 +6,7 @@
  * pgpool: a language independent connection pool server for PostgreSQL
  * written by Tatsuo Ishii
  *
- * Copyright (c) 2003-2015	PgPool Global Development Group
+ * Copyright (c) 2003-2021	PgPool Global Development Group
  *
  * Permission to use, copy, modify, and distribute this software and
  * its documentation for any purpose and without fee is hereby
@@ -59,7 +59,7 @@ extern void pool_init_process_context(void);
 extern POOL_PROCESS_CONTEXT * pool_get_process_context(void);
 extern ProcessInfo * pool_get_my_process_info(void);
 extern void pool_incremnet_local_session_id(void);
-extern int	pool_coninfo_size(void);
+extern size_t	pool_coninfo_size(void);
 extern int	pool_coninfo_num(void);
 extern ConnectionInfo * pool_coninfo(int child, int connection_pool, int backend);
 extern ConnectionInfo * pool_coninfo_pid(int pid, int connection_pool, int backend);
